+++ tracker-library/src/main/java/kaist/iclab/tracker/sensor/survey/activity/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package kaist.iclab.tracker.sensor.survey.activity.ui
+package kaist.iclab.tracker.sensor.survey.activity.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
